+++ packages/ui/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Patch Changes
 
 - Bump
+- Added defaultValue to field definitions
 
 ## 0.7.130